update .gitignore, update client with servables
